add(): add alerts in to actions notes

+++ src/app/pages/archive/archive.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { NotesService } from 'src/app/core/services/notes.service';
 import { NoteInterface } from 'src/app/shared/interfaces/note';
-
+/**
+ * Const Swal Defined
+ */
+ import Swal from "sweetalert2";
 @Component({
   selector: 'app-archive',
   templateUrl: './archive.component.html',
@@ -46,19 +49,44 @@ export class ArchiveComponent implements OnInit {
   * @description 
   * Update to Selected Note
   */
-  updateNote( $event: NoteInterface ): void {
-    this.noteService.updateNote($event).subscribe( );
+   updateNote( $event: NoteInterface ): void {
+    this.noteService.updateNote($event).subscribe( () => {
+      Swal.fire({
+        position: 'top-end',
+        icon: 'success',
+        title: 'The note has been updated',
+        showConfirmButton: false,
+        timer: 1500
+      })
+    } );
   }
   /**
-   * @memberof ArchiveComponent
+   * @memberof HomeComponent
    * @method deteteNote
    * @param { void }
    * @returns { void }
    * @description 
-   * Delete to Selected Note
+   * Delete to Selected Note and onfirm delete
    */
   deteteNote( $event: NoteInterface ): void {
-    this.noteService.deleteNote($event).subscribe( );
+    Swal.fire({
+      title: 'Are you sure to delete this note?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.noteService.deleteNote($event).subscribe(  );
+        Swal.fire(
+          'Deleted!',
+          'Your note has been deleted.',
+          'success'
+        )
+      }
+    });
   }
 
 }
